[Fix] 주문 생성 중 에러 수정 - #99

+++ src/main/java/com/nhnacademy/heukbaekfrontend/book/dto/response/BookResponse.java
@@ -4,13 +4,14 @@ import com.nhnacademy.heukbaekfrontend.contributor.dto.ContributorSummaryRespons
 import com.nhnacademy.heukbaekfrontend.image.dto.ImageSummaryResponse;
 import com.nhnacademy.heukbaekfrontend.publisher.dto.PublisherSummaryResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record BookResponse(Long id,
                            String title,
                            String publishedAt,
                            String salePrice,
-                           float discountRate,
+                           BigDecimal discountRate,
                            String thumbnailUrl,
                            List<ContributorSummaryResponse> contributors,
                            PublisherSummaryResponse publisher) {
